feat search

+++ src/bean/Painting.java
@@ -1,5 +1,7 @@
 package bean;
 
+import java.math.BigDecimal;
+
 public class Painting {
     private int paintingID;
     private String imageFileName;
@@ -11,7 +13,7 @@ public class Painting {
     private int width;
     private int height;
     private int cost;
-    private int msrp;
+    private BigDecimal msrp;
 
     public int getPaintingID() {
         return paintingID;
@@ -69,11 +71,11 @@ public class Painting {
         this.cost = cost;
     }
 
-    public int getMsrp() {
+    public BigDecimal getMsrp() {
         return msrp;
     }
 
-    public void setMsrp(int msrp) {
+    public void setMsrp(BigDecimal msrp) {
         this.msrp = msrp;
     }
